Fix postinstall scripts for a win32 environment (#25)
The postscript fails on win32 machine. And as [browser-compat-data](https://github.com/mdn/browser-compat-data) is dependent on this package the ```npm install``` fails for it too. This pull request aims to fix only postinstall script so that ```npm install``` can be fixed for win32 environment.

This is my first pull request, so please help me out in fixing anything that is not correct.