Switch the confluence dependency to GoogleChromeLabs/confluence
The mdn-confluence branch in this repo points to the same commit
(53d37216539948b094063d0f94bf7c986e6fb875) as mdn in the old.

The change of branch name is for consistency with foam2, and because the
mdn branch is used in https://github.com/GoogleChromeLabs/confluence/pull/344.

This will allow further changes on that branch.