Revamp how Confluence releases are processed/iterated
Strip out the `confluenceDAO` indirection around the data. Something in
this layer seems to have mixed up the order of releases, resulting in
spurious data. Just use the JSON files directly.

Because we have multiple releases in Confluence for some BCD releases,
group by BCD release and treat as supported if any of the Confluence
releases have support. We have this especially for macOS and iOS, with
multiple OS versions for the same Safari version.